Add enemy attack to main building

+++ Assets/_Project/Scripts/Descriptors/EnemyDescriptor.cs
@@ -7,7 +7,7 @@ namespace _Project.Scripts.Descriptors
     public class EnemyDescriptor : ScriptableObject
     {
         public Enemy Enemy;
-        public int Health;
+        public float Health;
         public int EnemiesNumber;
         public float MoveSpeed;
         public float PursuitDistance;
